Move reused code
