+++ src/test/kotlin/com/projectcitybuild/tests/integration/bans/CreateBanActionTests.kt
@@ -1,10 +1,10 @@
 package com.projectcitybuild.tests.integration.bans
 
-import com.projectcitybuild.core.network.NetworkClients
-import com.projectcitybuild.core.network.mojang.client.MojangClient
-import com.projectcitybuild.core.network.pcb.client.PCBClient
 import com.projectcitybuild.modules.bans.CreateBanAction
-import junit.framework.TestCase.*
+import com.projectcitybuild.tests.mocks.makeNetworkClients
+import com.projectcitybuild.tests.mocks.withJSONResource
+import junit.framework.TestCase.assertEquals
+import junit.framework.TestCase.fail
 import okhttp3.mockwebserver.MockResponse
 import okhttp3.mockwebserver.MockWebServer
 import org.hamcrest.CoreMatchers.instanceOf
@@ -12,7 +12,6 @@ import org.hamcrest.MatcherAssert.assertThat
 import org.junit.After
 import org.junit.Before
 import org.junit.Test
-import java.io.InputStreamReader
 import java.util.*
 
 class CreateBanActionTests {
@@ -30,42 +29,12 @@ class CreateBanActionTests {
         mockWebServer.shutdown()
     }
 
-    private fun jsonMockResponse(fileName: String): String {
-        val reader = InputStreamReader(javaClass.classLoader.getResourceAsStream(fileName))
-        val content = reader.readText()
-        reader.close()
-        return content
-    }
-
-    private fun networkClients(mockWebServer: MockWebServer): NetworkClients {
-        val baseUrl = mockWebServer.url("")
-        return NetworkClients(
-                PCBClient(
-                        authToken = "",
-                        withLogging = false,
-                        baseUrl = "$baseUrl/"
-                ),
-                MojangClient(
-                        withLogging = false
-                )
-        )
-    }
-
-    private fun mockResponse(fileName: String): MockResponse {
-        val json = jsonMockResponse(fileName)
-
-        return MockResponse()
-                .setBody(json)
-                .setHeader("Content-Type", "application/json")
-    }
-
     @Test
     fun `ban response returns success`() {
-        val response = mockResponse("api_ban_status_ban_response.json")
+        val response = MockResponse().withJSONResource("api_ban_status_ban_response.json")
         mockWebServer.enqueue(response)
 
-        val networkClients = networkClients(mockWebServer)
-        val action = CreateBanAction(networkClients)
+        val action = CreateBanAction(mockWebServer.makeNetworkClients())
         val result = action.execute(
                 playerId = UUID.randomUUID(),
                 playerName = "test_user",
@@ -77,11 +46,10 @@ class CreateBanActionTests {
 
     @Test
     fun `invalid input returns failure`() {
-        val response = mockResponse("api_ban_invalid_input_response.json")
+        val response = MockResponse().withJSONResource("api_ban_invalid_input_response.json")
         mockWebServer.enqueue(response)
 
-        val networkClients = networkClients(mockWebServer)
-        val action = CreateBanAction(networkClients)
+        val action = CreateBanAction(mockWebServer.makeNetworkClients())
         val result = action.execute(
                 playerId = UUID.randomUUID(),
                 playerName = "test_user",
@@ -96,11 +64,10 @@ class CreateBanActionTests {
 
     @Test
     fun `banning an already-banned player returns failure`() {
-        val response = mockResponse("api_ban_already_banned_response.json")
+        val response = MockResponse().withJSONResource("api_ban_already_banned_response.json")
         mockWebServer.enqueue(response)
 
-        val networkClients = networkClients(mockWebServer)
-        val action = CreateBanAction(networkClients)
+        val action = CreateBanAction(mockWebServer.makeNetworkClients())
         val result = action.execute(
                 playerId = UUID.randomUUID(),
                 playerName = "test_user",
